Move test client inline restarts to own file
